rec/requirements.txt
scikit-learn==1.5.0
numpy==1.26.4
face_recognition
opencv-contrib-python
setuptools
python-dotenv
gunicorn
eventlet
flask
flask-socketio
boto3
image
Flask-Cors